# Optimize building the communication map