--- a/docs/changelog/1830.md
+++ b/docs/changelog/1830.md
@@ -1,0 +1,1 @@
+- Optimized communication map creation for parallel single-level initialization significantly.