## Main changes of this PR

Optimizes the function [`buildCommunicationMap`](https://github.com/precice/precice/blob/7bdf93d6e060bad08cbef26946ea0536478251cd/src/m2n/PointToPointCommunication.cpp#L255) significantly.

## Motivation and additional information

While investigating PUM mappings, running larger cases on more cores became infeasible. After digging a bit deeper into the reason, it turned out that creating the 'lookup table' in `buildCommunicationMap` takes 99% of the time. So, I spent some effort to optimize the current way we compute the communication map. I'm not 100% sure that the proposed changes here are actually legal, as it performs a sorting on the vertex indices. However, I _could_ recover less, but still significant speed-up without doing so.

The changes are mostly relevant for a large number of global vertices. Here are some relevant performance measurements for one of the aste standard cases (one time-step, one mapping, serial-explicit schemes) on a 600k vertices input mesh and corresponding speedups on SuperMUC-NG

On 2 x 2 cores (input x output)

|   | `buildCommMap` | global run time |
| ------------- | ------------- | ------------- |
| `develop` | 2,081,519 |16,892,952 (~17secs) |
| `opt-bbmap` | 3,360  | 14,845,686 (~15secs) |
| speed up | 619 | 1.14 |

On 24 x 24 cores (input x output)

|   | `buildCommMap` | global run time |
| ------------- | ------------- | ------------- |
| `develop` |  839,366,952 |858,955,792 (~14min) |
| `opt-bbmap` | 2,358  | 4,554,192 (~4.5secs)|
| speed up | 355,965  | 188 |


On 240 x 240 cores (input x output)

|   | `buildCommMap` | global run time |
| ------------- | ------------- | ------------- |
| `develop` | 4,645,940,719 |4,688,943,494 (~77min)|
| `opt-bbmap` |  2,235 | 10,057,928 (~10 secs) |
| speed up |  2,078,720 | 466 |

Note that the issue is a general preCICE issue. However, using a NN mapping (as opposed to the PUM mapping used here) with a filter on the primary rank reduces the vertex count relevant for the function significantly (mapping fine to coarse), such that the issue would be observable only later. Also note that the global run time is now dominated by other events instead of the BB map creation. The larger the input mesh gets, the more significant the changes here would be (and vice versa).

EDIT: With the first version of this PR, cases with 1.35mio (bbMap in 3,300us) ran on 240 cores in <1min as well.

## Author's checklist

* [x] I used the [`pre-commit` hook](https://precice.org/dev-docs-dev-tooling.html#setting-up-pre-commit) to prevent dirty commits and used `pre-commit run --all` to format old commits.
* [ ] I added a changelog file with `make changelog` if there are user-observable changes since the last release.
* [ ] I added a test to cover the proposed changes in our test suite.
* [x] I sticked to C++17 features.
* [ ] I squashed / am about to squash all commits that should be seen as one.

## Reviewers' checklist

<!-- Tag people next to each point and add points for specific questions -->

* [ ] Does the changelog entry make sense? Is it formatted correctly?
* [ ] Do you understand the code changes?

<!-- add more questions/tasks if necessary -->