@fsimonis 2023-10-06
The following works on develop. This would be a good addition to this PR to prevent strange things from happening in the future:

```cpp
void setVertexDistribution(VertexDistribution vd)
{
  PRECICE_ASSERT(std::all_of(vd.begin(), vd.end(), [](const auto& p){return std::is_sorted(p.second.begin(),p.second.end());}));
  _vertexDistribution = std::move(vd);
}
```