@davidscn 2023-10-06
Assuming that the processed index ranges are sorted (I only tested the most costly case with 240 cores) gives an _additional_ speed-up of 4 - 7 (required 288 to 500 us). Actually, the timings are now more diverse since the total cost depends on how many intersections we have to go through on each rank whereas before (ie with the `std::is_sorted`) we had anyway a linear complexity on each rank. I'm sure with less ranks, the speed-up will be less though.
However, using this version we could directly remove the whole event, as it looks somewhat negligible.