@fsimonis 2023-10-06
> I'm sure with less ranks, the speed-up will be less though.
However, using this version we could directly remove the whole event, as it looks somewhat negligible.

This should be easy to check, as we still have the event. Given that this used to be a major bottleneck, I would keep the event. This allows to directly compare preCICE versions.

If we ever need to improve this further, then the next step is possibly to use ranges instead of individual ids.
So instead of storing $(0,...,100,200,...,10000)$, we store first+last $(0,100,200,10000)$ or first+count $(0,100,200,9800)$, making intersection handling etc way faster.